Rename PriorityTimeoutProperty to OccupyTimeoutProperty
